Changes coaches to practicioners
This aligns a lot better with the way we work and think. Bulding amazing
teams with happy develoeprs that respond to change is only able to be
done from the inside. A coach can only take you so far, but a coach who
is also a player can take you to the next level.

Amos King @adkron <amos@binarynoggin.com>